각 메뉴 enum 이용 bowling, lotto, home 각 클래스 파일로 분리
